Move image cache to integrations
