+++ tests/test_utils/test_image_cache.py
@@ -5,13 +5,13 @@ from unittest import mock
 import pytest
 from django.test import override_settings
 
+from tilavarauspalvelu.integrations import image_cache
 from tilavarauspalvelu.integrations.sentry import SentryLogger
-from utils import image_cache
 
 from tests.helpers import patch_method
 
 
-@mock.patch("utils.image_cache.urljoin")
+@mock.patch("tilavarauspalvelu.integrations.image_cache.urljoin")
 def test_image_cache_purge__no_action_if_disabled(urljoin):
     image_cache.purge("foo/bar.jpg")
     assert urljoin.called is False
@@ -35,7 +35,7 @@ def test_image_cache_purge__error_if_cache_purge_key_missing():  # NOSONAR pytho
 
 @override_settings(IMAGE_CACHE_ENABLED=True)
 @patch_method(SentryLogger.log_message)
-@mock.patch("utils.image_cache.request")
+@mock.patch("tilavarauspalvelu.integrations.image_cache.request")
 def test_image_cache_purge__makes_correct_request(request):
     request.return_value = mock.MagicMock(status_code=200)
     image_cache.purge("foo/bar.jpg")
@@ -50,7 +50,7 @@ def test_image_cache_purge__makes_correct_request(request):
 
 @override_settings(IMAGE_CACHE_ENABLED=True)
 @patch_method(SentryLogger.log_message)
-@mock.patch("utils.image_cache.request")
+@mock.patch("tilavarauspalvelu.integrations.image_cache.request")
 def test_image_cache_purge__logs_failed_requests(request):
     request.return_value = mock.MagicMock(status_code=400)
