+++ tilavarauspalvelu/integrations/image_cache.py
